requirements.txt
gradio
langchain
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
faiss-cpu
PyPDF2